requirements.txt
gunicorn
python-telegram-bot
pyowm